@@ -1,10 +1,10 @@
 module github.com/schwarzlichtbezirk/pds-grpc
 
-go 1.15
+go 1.16
 
 require (
 	github.com/gorilla/mux v1.8.0
-	google.golang.org/grpc v1.35.0
+	google.golang.org/grpc v1.36.0
 	google.golang.org/protobuf v1.25.0
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
