@@ -8,8 +8,10 @@ func main() {
 	log.Println("starts")
 	Run()
 	log.Printf("ready")
-	WaitBreak()
-	log.Println("shutting down begin")
-	Shutdown()
+	go func() {
+		WaitBreak()
+		log.Println("shutting down by break begin")
+	}()
+	WaitExit()
 	log.Println("shutting down complete.")
 }
